+++ spython/instance/cmd/start.py
@@ -18,7 +18,7 @@ def start(
     options=None,
     capture=False,
     singularity_options=None,
-    env=None,
+    environ=None,
 ):
     """start an instance. This is done by default when an instance is created.
 
@@ -80,7 +80,7 @@ def start(
     self.cmd = cmd
 
     output = run_command(
-        cmd, sudo=sudo, sudo_options=sudo_options, quiet=True, capture=capture, env=env
+        cmd, sudo=sudo, sudo_options=sudo_options, quiet=True, capture=capture, environ=environ
     )
 
     if output["return_code"] == 0: