+++ spython/utils/terminal.py
@@ -154,7 +154,7 @@ def run_command(
     no_newline_regexp="Progess",
     quiet=False,
     sudo_options=None,
-    env=None,
+    environ=None,
 ):
 
     """run_command uses subprocess to send a command to the terminal. If
@@ -181,7 +181,7 @@ def run_command(
 
     # Use the parent stdout and stderr
 
-    process = subprocess.Popen(cmd, stderr=subprocess.PIPE, stdout=stdout, env=env)
+    process = subprocess.Popen(cmd, stderr=subprocess.PIPE, stdout=stdout, env=environ)
     lines = []
     found_match = False